refactor: moved handleSubmit function to UserForm
