+++ src/pages/users/[name]/edit.tsx
@@ -1,43 +1,17 @@
-import { useMutation } from "@apollo/client";
 import { Card, CardContent, Typography } from "@mui/material";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
-import { UPDATE_USER_MUTATION } from "../../../client/users/mutations";
 import ProgressBar from "../../../components/Shared/ProgressBar";
 import UserForm from "../../../components/Users/Form";
 import { useTranslate } from "../../../hooks/common";
 import { useUserByNameQuery } from "../../../hooks/user";
-import { UpdateUserMutation, UserFormValues } from "../../../types/user";
-import { redirectTo } from "../../../utils/common";
-import { getUserProfilePath } from "../../../utils/user";
 
 const EditUser: NextPage = () => {
   const { query } = useRouter();
   const userName = String(query?.name || "");
   const [user, userLoading, error] = useUserByNameQuery(userName);
-  const [updateUser] = useMutation<UpdateUserMutation>(UPDATE_USER_MUTATION);
 
   const t = useTranslate();
-
-  const handleSubmit = async (formValues: UserFormValues) => {
-    try {
-      const { data } = await updateUser({
-        variables: {
-          userData: {
-            id: user?.id,
-            ...formValues,
-          },
-        },
-      });
-      if (!data) {
-        throw Error(t("errors.somethingWentWrong"));
-      }
-      const path = getUserProfilePath(data.updateUser.name);
-      redirectTo(path);
-    } catch (err) {
-      console.error(err);
-    }
-  };
 
   if (error) {
     return <Typography>{t("errors.somethingWentWrong")}</Typography>;
@@ -51,18 +25,11 @@ const EditUser: NextPage = () => {
     return null;
   }
 
-  const initialValues: UserFormValues = {
-    bio: user.bio || "",
-    email: user.email,
-    name: user.name,
-  };
-
   return (
     <Card>
       <CardContent>
         <UserForm
-          onSubmit={handleSubmit}
-          initialValues={initialValues}
+          editUser={user}
           submitButtonText={t("actions.save")}
           isEditing
         />
